+++ src/app/coreFilesUpdater/CoreFilesUpdaterFactory.ts
@@ -1,0 +1,17 @@
+// Copyright (c) Wictor Wilén. All rights reserved. 
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
+import { BaseCoreFilesUpdater } from "./BaseCoreFilesUpdater";
+import * as semver from "semver";
+import { CoreFilesUpdater_2_11 } from "./coreFilesUpdaters/CoreFilesUpdater_2_11";
+
+export class CoreFilesUpdaterFactory {
+    public static createCoreFilesUpdater(currentVersion: string): BaseCoreFilesUpdater | undefined {
+        // gulp.config.js update (2.9.* or 2.10.*)
+        if (semver.lt(currentVersion, "2.11.0")) {
+            return new CoreFilesUpdater_2_11();
+        }
+        return undefined;
+    }
+}