+++ src/localization/index.ts
@@ -1,0 +1,7 @@
+// Copyright (c) Wictor Wilén. All rights reserved. 
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
+import { LocalizationGenerator } from './LocalizationGenerator';
+
+module.exports = LocalizationGenerator